abstract argument parser into it's own library
