+++ lib/argument_parser/src/argument_keys.dart
@@ -1,0 +1,7 @@
+const from = 'from';
+const to = 'to';
+const psalmConfig = 'psalm-config';
+const frequency = 'frequency';
+const psalmVersion = 'psalm-version';
+const considerAllCommits = 'consider-all-commits';
+const help = 'help';
